AS<-3548 ASM has trouble configuring chassis either after a factory reset, or if BMP mode enabled
After restoring the switch to factory default, MXL switch can be in BMP mode, where commands are not accepted by the switch.
Added command sequence to come of the the mode and disable the BMP mode for further reboots. This adds ~15 seconds to the existing flow, but its required to take care of the BMP mode.